requirements.txt
et_xmlfile==2.0.0
lxml==6.0.2
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
six==1.17.0
typing_extensions==4.15.0
